fix: mobile menu props
